requirements.txt
Pillow==9.4.0
tqdm==4.64.1
